Add server and change dir structure
