+++ .gitignore
@@ -1,9 +1,8 @@
 /node_modules
 /package-lock.json
 
+/core
+/html
+/server
+/shared
 /coverage
-/.tsbuildinfo
-/__tests__
-/lib
-/core.*
-/index.*
