+++ package.json
@@ -2,27 +2,26 @@
   "name": "fuco",
   "version": "1.2.0",
   "description": "Functional Component like React, but for Web Components",
-  "main": "index.js",
-  "module": "index.js",
-  "typings": "index.d.ts",
+  "main": "core/fuco.js",
+  "module": "core/fuco.mjs",
+  "typings": "core/index.d.ts",
   "files": [
-    "lib",
-    "core.js",
-    "core.d.ts",
-    "core.js.map",
-    "index.d.ts",
-    "index.js",
-    "index.js.map"
+    "core",
+    "html",
+    "server",
+    "shared",
+    "!**/__tests__/**"
   ],
   "scripts": {
-    "build": "tsc",
-    "test": "karma start",
+    "build": "rollup -c",
+    "test": "npm run -s test:mocha && npm run -s test:karma",
+    "test:mocha": "TS_NODE_PROJECT=tsconfig.test.json mocha -r ts-node/register src/**/__tests__/**/*.spec.ts",
+    "test:karma": "karma start",
     "lint": "tsc --noEmit && eslint src --ext .ts",
-    "fmt": "npm run lint -- --fix",
-    "clean": "rm -fr coverage .tsbuildinfo __tests__ lib index.* core.*",
+    "fmt": "prettier --write 'src/**/*'",
+    "clean": "rm -fr core html server shared coverage",
     "distclean": "npm run clean && rm -fr node_modules package-lock.json",
-    "prepare": "npm run clean && npm run build",
-    "sizecheck": "npm run build && rollup -c"
+    "prepare": "npm run clean && npm run build"
   },
   "keywords": [
     "webcomponents",
@@ -34,28 +33,30 @@
   "homepage": "https://github.com/wtnbass/fuco",
   "repository": "wtnbass/fuco",
   "devDependencies": {
-    "@types/jasmine": "3.4.4",
-    "@typescript-eslint/eslint-plugin": "2.4.0",
-    "@typescript-eslint/parser": "2.4.0",
+    "@types/chai": "4.2.7",
+    "@types/mocha": "5.2.7",
+    "@types/sinon": "7.5.1",
+    "@typescript-eslint/eslint-plugin": "2.13.0",
+    "@typescript-eslint/parser": "2.13.0",
     "codecov": "3.6.1",
-    "eslint": "6.5.1",
-    "eslint-config-prettier": "6.4.0",
-    "eslint-plugin-prettier": "3.1.1",
+    "eslint": "6.8.0",
+    "eslint-config-prettier": "6.9.0",
+    "eslint-plugin-prettier": "3.1.2",
     "istanbul-instrumenter-loader": "3.0.1",
     "karma": "4.4.1",
     "karma-chrome-launcher": "3.1.0",
-    "karma-coverage-istanbul-reporter": "2.1.0",
+    "karma-coverage-istanbul-reporter": "2.1.1",
     "karma-firefox-launcher": "1.2.0",
     "karma-jasmine": "2.0.1",
     "karma-webpack": "4.0.2",
-    "prettier": "1.18.2",
-    "puppeteer": "1.20.0",
-    "rollup": "1.25.0",
+    "prettier": "1.19.1",
+    "rollup": "1.27.14",
     "rollup-plugin-filesize": "6.2.1",
-    "rollup-plugin-node-resolve": "5.2.0",
-    "rollup-plugin-terser": "5.1.2",
-    "ts-loader": "6.2.0",
-    "typescript": "3.6.4",
-    "webpack": "4.41.2"
+    "rollup-plugin-terser": "5.1.3",
+    "rollup-plugin-typescript2": "0.25.3",
+    "ts-loader": "6.2.1",
+    "ts-node": "8.5.4",
+    "typescript": "3.7.4",
+    "webpack": "4.41.5"
   }
 }
