--- src/core.ts
@@ -1,18 +1,0 @@
-export { html, render, HtmlTemplate } from "./lib/html";
-export { Component } from "./lib/component";
-export { css, unsafeCSS } from "./lib/css";
-export { createContext } from "./lib/context";
-export {
-  useAttribute,
-  useProperty,
-  useDispatchEvent,
-  useStyle,
-  useRef,
-  useState,
-  useReducer,
-  useContext,
-  useEffect,
-  useLayoutEffect,
-  useMemo,
-  useCallback
-} from "./lib/hooks";
