+++ src/core/__tests__/attribute_test.ts
