+++ src/core/__tests__/component_test.ts
