+++ src/core/__tests__/dispatch-event_test.ts
