+++ src/core/__tests__/fixture.ts
