+++ src/core/__tests__/style_test.ts
