+++ src/server/index.ts
@@ -1,0 +1,5 @@
+import { stringify } from "./stringify";
+
+export function renderToString(template: unknown) {
+  return stringify(1, [0, template]);
+}
