+++ tsconfig.test.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "target": "es2017",
-    "module": "es2015",
+    "module": "commonjs",
     "lib": ["es2017", "dom", "dom.iterable"],
     "moduleResolution": "node",
     "outDir": "./",
